Move all classes and objects that is not Object out of object.h/cpp
