+++ core/object/object.h
@@ -33,10 +33,10 @@
 #include "core/extension/gdextension_interface.gen.h"
 #include "core/object/gdtype.h"
 #include "core/object/method_info.h"
+#include "core/object/object_gdextension.h"
 #include "core/object/object_id.h"
 #include "core/object/property_info.h"
 #include "core/os/mutex.h"
-#include "core/os/spin_lock.h"
 #include "core/templates/hash_map.h"
 #include "core/templates/hash_set.h"
 #include "core/templates/list.h"
@@ -70,72 +70,6 @@
 // API used to extend in GDExtension and other C compatible compiled languages.
 class MethodBind;
 class GDExtension;
-
-struct ObjectGDExtension {
-	GDExtension *library = nullptr;
-	ObjectGDExtension *parent = nullptr;
-	List<ObjectGDExtension *> children;
-	StringName parent_class_name;
-	StringName class_name;
-	bool editor_class = false;
-	bool reloadable = false;
-	bool is_virtual = false;
-	bool is_abstract = false;
-	bool is_exposed = true;
-#ifdef TOOLS_ENABLED
-	bool is_runtime = false;
-	bool is_placeholder = false;
-#endif
-#ifndef DISABLE_DEPRECATED
-	bool legacy_unexposed_class = false;
-#endif // DISABLE_DEPRECATED
-	GDExtensionClassSet set;
-	GDExtensionClassGet get;
-	GDExtensionClassGetPropertyList get_property_list;
-	GDExtensionClassFreePropertyList2 free_property_list2;
-	GDExtensionClassPropertyCanRevert property_can_revert;
-	GDExtensionClassPropertyGetRevert property_get_revert;
-	GDExtensionClassValidateProperty validate_property;
-#ifndef DISABLE_DEPRECATED
-	GDExtensionClassNotification notification;
-	GDExtensionClassFreePropertyList free_property_list;
-#endif // DISABLE_DEPRECATED
-	GDExtensionClassNotification2 notification2;
-	GDExtensionClassToString to_string;
-	GDExtensionClassReference reference;
-	GDExtensionClassReference unreference;
-	GDExtensionClassGetRID get_rid;
-
-	void *class_userdata = nullptr;
-
-#ifndef DISABLE_DEPRECATED
-	GDExtensionClassCreateInstance create_instance;
-#endif // DISABLE_DEPRECATED
-	GDExtensionClassCreateInstance2 create_instance2;
-	GDExtensionClassFreeInstance free_instance;
-#ifndef DISABLE_DEPRECATED
-	GDExtensionClassGetVirtual get_virtual;
-	GDExtensionClassGetVirtualCallData get_virtual_call_data;
-#endif // DISABLE_DEPRECATED
-	GDExtensionClassGetVirtual2 get_virtual2;
-	GDExtensionClassGetVirtualCallData2 get_virtual_call_data2;
-	GDExtensionClassCallVirtualWithData call_virtual_with_data;
-	GDExtensionClassRecreateInstance recreate_instance;
-
-#ifdef TOOLS_ENABLED
-	void *tracking_userdata = nullptr;
-	void (*track_instance)(void *p_userdata, void *p_instance) = nullptr;
-	void (*untrack_instance)(void *p_userdata, void *p_instance) = nullptr;
-#endif
-
-	/// A type for this Object extension.
-	/// This is not exposed through the GDExtension API (yet) so it is inferred from above parameters.
-	GDType *gdtype;
-	void create_gdtype();
-	void destroy_gdtype();
-
-	~ObjectGDExtension();
-};
 
 #define GDVIRTUAL_CALL(m_name, ...) _gdvirtual_##m_name##_call(__VA_ARGS__)
 #define GDVIRTUAL_CALL_PTR(m_obj, m_name, ...) m_obj->_gdvirtual_##m_name##_call(__VA_ARGS__)
@@ -865,277 +799,6 @@ bool Object::derives_from() const {
 		}
 	}
 }
-
-class ObjectDB {
-// This needs to add up to 63, 1 bit is for reference.
-#define OBJECTDB_VALIDATOR_BITS 39
-#define OBJECTDB_VALIDATOR_MASK ((uint64_t(1) << OBJECTDB_VALIDATOR_BITS) - 1)
-#define OBJECTDB_SLOT_MAX_COUNT_BITS 24
-#define OBJECTDB_SLOT_MAX_COUNT_MASK ((uint64_t(1) << OBJECTDB_SLOT_MAX_COUNT_BITS) - 1)
-#define OBJECTDB_REFERENCE_BIT (uint64_t(1) << (OBJECTDB_SLOT_MAX_COUNT_BITS + OBJECTDB_VALIDATOR_BITS))
-
-	struct ObjectSlot { // 128 bits per slot.
-		uint64_t validator : OBJECTDB_VALIDATOR_BITS;
-		uint64_t next_free : OBJECTDB_SLOT_MAX_COUNT_BITS;
-		uint64_t is_ref_counted : 1;
-		Object *object = nullptr;
-	};
-
-	static SpinLock spin_lock;
-	static uint32_t slot_count;
-	static uint32_t slot_max;
-	static ObjectSlot *object_slots;
-	static uint64_t validator_counter;
-
-	friend class Object;
-	friend void unregister_core_types();
-	static void cleanup();
-
-	static ObjectID add_instance(Object *p_object);
-	static void remove_instance(Object *p_object);
-
-	friend void register_core_types();
-	static void setup();
-
-public:
-	typedef void (*DebugFunc)(Object *p_obj, void *p_user_data);
-
-	_ALWAYS_INLINE_ static Object *get_instance(ObjectID p_instance_id) {
-		uint64_t id = p_instance_id;
-		uint32_t slot = id & OBJECTDB_SLOT_MAX_COUNT_MASK;
-
-		ERR_FAIL_COND_V(slot >= slot_max, nullptr); // This should never happen unless RID is corrupted.
-
-		spin_lock.lock();
-
-		uint64_t validator = (id >> OBJECTDB_SLOT_MAX_COUNT_BITS) & OBJECTDB_VALIDATOR_MASK;
-
-		if (unlikely(object_slots[slot].validator != validator)) {
-			spin_lock.unlock();
-			return nullptr;
-		}
-
-		Object *object = object_slots[slot].object;
-
-		spin_lock.unlock();
-
-		return object;
-	}
-
-	template <typename T>
-	_ALWAYS_INLINE_ static T *get_instance(ObjectID p_instance_id) {
-		return Object::cast_to<T>(get_instance(p_instance_id));
-	}
-
-	template <typename T>
-	_ALWAYS_INLINE_ static Ref<T> get_ref(ObjectID p_instance_id); // Defined in ref_counted.h
-
-	static void debug_objects(DebugFunc p_func, void *p_user_data);
-	static int get_object_count();
-};
-
-// Using `RequiredResult<T>` as the return type indicates that null will only be returned in the case of an error.
-// This allows GDExtension language bindings to use the appropriate error handling mechanism for that language
-// when null is returned (for example, throwing an exception), rather than simply returning the value.
-template <typename T>
-class RequiredResult {
-	static_assert(!is_fully_defined_v<T> || std::is_base_of_v<Object, T>, "T must be an Object subtype");
-
-public:
-	using element_type = T;
-	using ptr_type = std::conditional_t<std::is_base_of_v<RefCounted, T>, Ref<T>, T *>;
-
-private:
-	ptr_type _value = ptr_type();
-
-public:
-	_FORCE_INLINE_ RequiredResult() = default;
-
-	RequiredResult(const RequiredResult &p_other) = default;
-	RequiredResult(RequiredResult &&p_other) = default;
-	RequiredResult &operator=(const RequiredResult &p_other) = default;
-	RequiredResult &operator=(RequiredResult &&p_other) = default;
-
-	_FORCE_INLINE_ RequiredResult(std::nullptr_t) :
-			RequiredResult() {}
-	_FORCE_INLINE_ RequiredResult &operator=(std::nullptr_t) { _value = nullptr; }
-
-	// These functions should not be called directly, they are only for internal use.
-	_FORCE_INLINE_ ptr_type _internal_ptr_dont_use() const { return _value; }
-	_FORCE_INLINE_ static RequiredResult<T> _err_return_dont_use() { return RequiredResult<T>(); }
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(const RequiredResult<T_Other> &p_other) :
-			_value(p_other._value) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(const RequiredResult<T_Other> &p_other) {
-		_value = p_other._value;
-		return *this;
-	}
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(T_Other *p_ptr) :
-			_value(p_ptr) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(T_Other *p_ptr) {
-		_value = p_ptr;
-		return *this;
-	}
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(const Ref<T_Other> &p_ref) :
-			_value(p_ref) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(const Ref<T_Other> &p_ref) {
-		_value = p_ref;
-		return *this;
-	}
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(Ref<T_Other> &&p_ref) :
-			_value(std::move(p_ref)) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(Ref<T_Other> &&p_ref) {
-		_value = std::move(p_ref);
-		return &this;
-	}
-
-	template <typename U = T, std::enable_if_t<std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(const Variant &p_variant) :
-			_value(Object::cast_to<T>(p_variant.get_validated_object())) {}
-	template <typename U = T, std::enable_if_t<std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(const Variant &p_variant) {
-		_value = Object::cast_to<T>(p_variant.get_validated_object());
-		return *this;
-	}
-
-	template <typename U = T, std::enable_if_t<!std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredResult(const Variant &p_variant) :
-			_value(Object::cast_to<T>(p_variant.operator Object *())) {}
-	template <typename U = T, std::enable_if_t<!std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredResult &operator=(const Variant &p_variant) {
-		_value = Object::cast_to<T>(p_variant.operator Object *());
-		return *this;
-	}
-
-	template <typename U = T, std::enable_if_t<std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ element_type *ptr() const {
-		return *_value;
-	}
-
-	template <typename U = T, std::enable_if_t<!std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ element_type *ptr() const {
-		return _value;
-	}
-
-	_FORCE_INLINE_ operator ptr_type() const {
-		return _value;
-	}
-
-	template <typename U = T, typename T_Other, std::enable_if_t<std::is_base_of_v<RefCounted, U> && std::is_base_of_v<U, T_Other>, int> = 0>
-	_FORCE_INLINE_ operator Ref<T_Other>() const {
-		return Ref<T_Other>(_value);
-	}
-
-	_FORCE_INLINE_ element_type *operator*() const {
-		return ptr();
-	}
-
-	_FORCE_INLINE_ element_type *operator->() const {
-		return ptr();
-	}
-};
-
-// Using `RequiredParam<T>` as an argument type indicates that passing null as that parameter is an error,
-// that will prevent the method from doing its intended function.
-// This allows GDExtension bindings to use language-specific mechanisms to prevent users from passing null,
-// because it is never valid to do so.
-template <typename T>
-class RequiredParam {
-	static_assert(!is_fully_defined_v<T> || std::is_base_of_v<Object, T>, "T must be an Object subtype");
-
-public:
-	static constexpr bool is_ref = std::is_base_of_v<RefCounted, T>;
-
-	using element_type = T;
-	using extracted_type = std::conditional_t<is_ref, const Ref<T> &, T *>;
-	using persistent_type = std::conditional_t<is_ref, Ref<T>, T *>;
-
-private:
-	T *_value = nullptr;
-
-	_FORCE_INLINE_ RequiredParam() = default;
-
-public:
-	// These functions should not be called directly, they are only for internal use.
-	_FORCE_INLINE_ extracted_type _internal_ptr_dont_use() const {
-		if constexpr (is_ref) {
-			// Pretend _value is a Ref, for ease of use with existing `const Ref &` accepting APIs.
-			// This only works as long as Ref is internally T *.
-			// The double indirection should be optimized away by the compiler.
-			static_assert(sizeof(Ref<T>) == sizeof(T *));
-			return *((const Ref<T> *)&_value);
-		} else {
-			return _value;
-		}
-	}
-	_FORCE_INLINE_ bool _is_null_dont_use() const { return _value == nullptr; }
-	_FORCE_INLINE_ static RequiredParam<T> _err_return_dont_use() { return RequiredParam<T>(); }
-
-	// Prevent erroneously assigning null values by explicitly removing nullptr constructor/assignment.
-	RequiredParam(std::nullptr_t) = delete;
-	RequiredParam &operator=(std::nullptr_t) = delete;
-
-	RequiredParam(const RequiredParam &p_other) = default;
-	RequiredParam(RequiredParam &&p_other) = default;
-	RequiredParam &operator=(const RequiredParam &p_other) = default;
-	RequiredParam &operator=(RequiredParam &&p_other) = default;
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam(const RequiredParam<T_Other> &p_other) :
-			_value(p_other._internal_ptr_dont_use()) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam &operator=(const RequiredParam<T_Other> &p_other) {
-		_value = p_other._internal_ptr_dont_use();
-		return *this;
-	}
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam(T_Other *p_ptr) :
-			_value(p_ptr) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam &operator=(T_Other *p_ptr) {
-		_value = p_ptr;
-		return *this;
-	}
-
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam(const Ref<T_Other> &p_ref) :
-			_value(*p_ref) {}
-	template <typename T_Other, std::enable_if_t<std::is_base_of_v<T, T_Other>, int> = 0>
-	_FORCE_INLINE_ RequiredParam &operator=(const Ref<T_Other> &p_ref) {
-		_value = *p_ref;
-		return *this;
-	}
-
-	template <typename U = T, std::enable_if_t<std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredParam(const Variant &p_variant) :
-			_value(Object::cast_to<T>(p_variant.get_validated_object())) {}
-	template <typename U = T, std::enable_if_t<std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredParam &operator=(const Variant &p_variant) {
-		_value = Object::cast_to<T>(p_variant.get_validated_object());
-		return *this;
-	}
-
-	template <typename U = T, std::enable_if_t<!std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredParam(const Variant &p_variant) :
-			_value(Object::cast_to<T>(p_variant.operator Object *())) {}
-	template <typename U = T, std::enable_if_t<!std::is_base_of_v<RefCounted, U>, int> = 0>
-	_FORCE_INLINE_ RequiredParam &operator=(const Variant &p_variant) {
-		_value = Object::cast_to<T>(p_variant.operator Object *());
-		return *this;
-	}
-};
 
 #define TMPL_EXTRACT_PARAM_OR_FAIL(m_name, m_param, m_retval, m_msg, m_editor) \
 	if (unlikely(m_param._is_null_dont_use())) { \
