+++ core/variant/method_ptrcall.h
@@ -30,8 +30,8 @@
 
 #pragma once
 
-#include "core/object/object.h"
 #include "core/object/object_id.h"
+#include "core/object/required_result.h"
 #include "core/templates/simple_type.h"
 #include "core/typedefs.h"
 #include "core/variant/variant.h"
