+++ core/variant/type_info.h
@@ -30,7 +30,7 @@
 
 #pragma once
 
-#include "core/object/object.h"
+#include "core/object/required_result.h"
 #include "core/templates/simple_type.h"
 #include "core/typedefs.h"
 #include "core/variant/variant.h"
